@@ -1,7 +1,6 @@
 <?php
 /**
  * CodeIgniter_Sniffs_Strings_DoubleQuoteUsageSniff.
- *
  * PHP version 5
  *
  * @category  PHP
@@ -14,7 +13,6 @@
 
 /**
  * CodeIgniter_Sniffs_Strings_DoubleQuoteUsageSniff.
- *
  * Ensures that double-quoted strings are used only to parse variables,
  * to avoid escape characters before single quotes or for chars that need
  * to be interpreted like \r, \n or \t.
@@ -32,11 +30,11 @@
 
 namespace CodeIgniter\Sniffs\Strings;
 
-use PHP_CodeSniffer\Sniffs\Sniff;
 use PHP_CodeSniffer\Files\File;
 
 class DoubleQuoteUsageSniff extends VariableUsageSniff
 {
+
     /**
      * Returns an array of tokens this test wants to listen for.
      *
@@ -44,17 +42,17 @@ class DoubleQuoteUsageSniff extends VariableUsageSniff
      */
     public function register()
     {
-        return array(
+        return [
             T_DOUBLE_QUOTED_STRING,
             T_CONSTANT_ENCAPSED_STRING,
-        );
+        ];
     }//end register()
 
     /**
      * Processes this test, when one of its tokens is encountered.
      *
-     * @param File $phpcsFile The current file being scanned.
-     * @param int                  $stackPtr  The position of the current token
+     * @param File $phpcsFile                 The current file being scanned.
+     * @param int  $stackPtr                  The position of the current token
      *                                        in the stack passed in $tokens.
      *
      * @return void
@@ -62,7 +60,7 @@ class DoubleQuoteUsageSniff extends VariableUsageSniff
     public function process(File $phpcsFile, $stackPtr)
     {
         // no variable are in the string from here
-        $tokens = $phpcsFile->getTokens();
+        $tokens   = $phpcsFile->getTokens();
         $qtString = $tokens[$stackPtr]['content'];
         // makes sure that it is about a double quote string,
         // since variables are not parsed out of double quoted string
@@ -73,102 +71,101 @@ class DoubleQuoteUsageSniff extends VariableUsageSniff
 
         if (0 === strcmp($open_qt_str, '"')) {
             $this->processDoubleQuotedString($phpcsFile, $stackPtr, $qtString);
-        } else if (0 === strcmp($open_qt_str, "'")) {
-            $this->processSingleQuotedString($phpcsFile, $stackPtr, $qtString);
+        } else {
+            if (0 === strcmp($open_qt_str, "'")) {
+                $this->processSingleQuotedString($phpcsFile, $stackPtr, $qtString);
+            }
         }
     }//end process()
-
 
     /**
      * Processes this test, when the token encountered is a double-quoted string.
      *
-     * @param File $phpcsFile The current file being scanned.
-     * @param int                  $stackPtr  The position of the current token
+     * @param File   $phpcsFile               The current file being scanned.
+     * @param int    $stackPtr                The position of the current token
      *                                        in the stack passed in $tokens.
-     * @param string               $qtString  The double-quoted string content,
+     * @param string $qtString                The double-quoted string content,
      *                                        i.e. without quotes.
      *
      * @return void
      */
-    protected function processDoubleQuotedString (File $phpcsFile, $stackPtr, $qtString)
+    protected function processDoubleQuotedString(File $phpcsFile, $stackPtr, $qtString)
     {
         // so there should be at least a single quote or a special char
         // if there are the 2 kinds of quote and no special char, then add a warning
-        $has_variable = parent::processDoubleQuotedString($phpcsFile, $stackPtr, '"'.$qtString.'"');
+        $has_variable          = parent::processDoubleQuotedString($phpcsFile, $stackPtr, '"' . $qtString . '"');
         $has_specific_sequence = $this->_hasSpecificSequence($qtString);
-        $dbl_qt_at = strpos($qtString, '"');
-        $smpl_qt_at = strpos($qtString, "'");
+        $dbl_qt_at             = strpos($qtString, '"');
+        $smpl_qt_at            = strpos($qtString, "'");
         if (false === $has_variable && false === $has_specific_sequence
             && false === $smpl_qt_at
         ) {
             $error = 'Single-quoted strings should be used unless it contains variables, special chars like \n or single quotes.';
             $phpcsFile->addError($error, $stackPtr, '');
-        } else if (false !== $smpl_qt_at && false !== $dbl_qt_at
-            && false === $has_variable && false === $has_specific_sequence
-        ) {
-            $warning = 'It is encouraged to use a single-quoted string, since it doesn\'t contain any variable nor special char though it mixes single and double quotes.';
-            $phpcsFile->addWarning($warning, $stackPtr, '');
+        } else {
+            if (false !== $smpl_qt_at && false !== $dbl_qt_at
+                && false === $has_variable && false === $has_specific_sequence
+            ) {
+                $warning = 'It is encouraged to use a single-quoted string, since it doesn\'t contain any variable nor special char though it mixes single and double quotes.';
+                $phpcsFile->addWarning($warning, $stackPtr, '');
+            }
         }
     }//end processDoubleQuotedString()
-
-
-    /**
-     * Processes this test, when the token encountered is a single-quoted string.
-     *
-     * @param File $phpcsFile The current file being scanned.
-     * @param int                  $stackPtr  The position of the current token
-     *                                        in the stack passed in $tokens.
-     * @param string               $qtString  The single-quoted string content,
-     *                                        i.e. without quotes.
-     *
-     * @return void
-     */
-    protected function processSingleQuotedString (File $phpcsFile, $stackPtr, $qtString)
-    {
-        // if there is single quotes without additional double quotes,
-        // then user is allowed to use double quote to avoid having to
-        // escape single quotes. Don't add the warning, if an error was
-        // already added, because a variable was found in a single-quoted
-        // string.
-        $has_variable = parent::processSingleQuotedString($phpcsFile, $stackPtr, "'".$qtString."'");
-        $dbl_qt_at = strpos($qtString, '"');
-        $smpl_qt_at = strpos($qtString, "'");
-        if (false === $has_variable && false !== $smpl_qt_at && false === $dbl_qt_at) {
-            $warning = 'You may also use double-quoted strings if the string contains single quotes, so you do not have to use escape characters.';
-            $phpcsFile->addWarning($warning, $stackPtr, '');
-        }
-    }//end processSingleQuotedString()
 
     /**
      * Return TRUE, if a sequence of chars that is parsed in a specific way
      * in double-quoted strings is found, FALSE otherwise.
      *
      * @param string $string String in which sequence of special chars will
-     * be researched.
+     *                       be researched.
      *
      * @return TRUE, if a sequence of chars that is parsed in a specific way
      * in double-quoted strings is found, FALSE otherwise.
-     *
      * @link http://www.php.net/manual/en/language.types.string.php#language.types.string.syntax.double
      */
     private function _hasSpecificSequence($string)
     {
-        $hasSpecificSequence = FALSE;
-        $specialMeaningStrs = array('\n', '\r', '\t', '\v', '\f');
+        $hasSpecificSequence = false;
+        $specialMeaningStrs  = ['\n', '\r', '\t', '\v', '\f'];
         foreach ($specialMeaningStrs as $splStr) {
-            if (FALSE !== strpos($string, $splStr)) {
-                $hasSpecificSequence = TRUE;
+            if (false !== strpos($string, $splStr)) {
+                $hasSpecificSequence = true;
             }
         }
-        $specialMeaningPtrns = array('\[0-7]{1,3}', '\x[0-9A-Fa-f]{1,2}');
+        $specialMeaningPtrns = ['\[0-7]{1,3}', '\x[0-9A-Fa-f]{1,2}'];
         foreach ($specialMeaningPtrns as $splPtrn) {
             if (1 === preg_match("/{$splPtrn}/", $string)) {
-                $hasSpecificSequence = TRUE;
+                $hasSpecificSequence = true;
             }
         }
+
         return $hasSpecificSequence;
+    }//end processSingleQuotedString()
+
+    /**
+     * Processes this test, when the token encountered is a single-quoted string.
+     *
+     * @param File   $phpcsFile               The current file being scanned.
+     * @param int    $stackPtr                The position of the current token
+     *                                        in the stack passed in $tokens.
+     * @param string $qtString                The single-quoted string content,
+     *                                        i.e. without quotes.
+     *
+     * @return void
+     */
+    protected function processSingleQuotedString(File $phpcsFile, $stackPtr, $qtString)
+    {
+        // if there is single quotes without additional double quotes,
+        // then user is allowed to use double quote to avoid having to
+        // escape single quotes. Don't add the warning, if an error was
+        // already added, because a variable was found in a single-quoted
+        // string.
+        $has_variable = parent::processSingleQuotedString($phpcsFile, $stackPtr, "'" . $qtString . "'");
+        $dbl_qt_at    = strpos($qtString, '"');
+        $smpl_qt_at   = strpos($qtString, "'");
+        if (false === $has_variable && false !== $smpl_qt_at && false === $dbl_qt_at) {
+            $warning = 'You may also use double-quoted strings if the string contains single quotes, so you do not have to use escape characters.';
+            $phpcsFile->addWarning($warning, $stackPtr, '');
+        }
     }//end _hasSpecificSequence()
-
 }//end class
-
-?>
