@@ -1,7 +1,6 @@
 <?php
 /**
  * CodeIgniter_Sniffs_Files_Utf8EncodingSniff.
- *
  * PHP version 5
  *
  * @category  PHP
@@ -14,7 +13,6 @@
 
 /**
  * CodeIgniter_Sniffs_Files_Utf8EncodingSniff.
- *
  * Ensures that PHP files are encoded with Unicode (UTF-8) encoding.
  *
  * @category  PHP
@@ -27,8 +25,8 @@
 
 namespace CodeIgniter\Sniffs\Files;
 
-use PHP_CodeSniffer\Sniffs\Sniff;
 use PHP_CodeSniffer\Files\File;
+use PHP_CodeSniffer\Sniffs\Sniff;
 
 class Utf8EncodingSniff implements Sniff
 {
@@ -40,18 +38,16 @@ class Utf8EncodingSniff implements Sniff
      */
     public function register()
     {
-        return array(
-            T_OPEN_TAG
-        );
-
+        return [
+            T_OPEN_TAG,
+        ];
     }//end register()
-
 
     /**
      * Processes this test, when one of its tokens is encountered.
      *
-     * @param File $phpcsFile The current file being scanned.
-     * @param int                  $stackPtr  The position of the current token
+     * @param File $phpcsFile                 The current file being scanned.
+     * @param int  $stackPtr                  The position of the current token
      *                                        in the stack passed in $tokens.
      *
      * @return void
@@ -65,23 +61,22 @@ class Utf8EncodingSniff implements Sniff
             }
         }
 
-        $file_path = $phpcsFile->getFilename();
-        $file_name = basename($file_path);
+        $file_path    = $phpcsFile->getFilename();
+        $file_name    = basename($file_path);
         $file_content = file_get_contents($file_path);
         if (false === mb_check_encoding($file_content, 'UTF-8')) {
             $error = 'File "' . $file_name . '" should be saved with Unicode (UTF-8) encoding.';
             $phpcsFile->addError($error, 0, '');
         }
-        if ( ! self::_checkUtf8W3c($file_content)) {
+        if (!self::_checkUtf8W3c($file_content)) {
             $error = 'File "' . $file_name . '" should be saved with Unicode (UTF-8) encoding, but it did not successfully pass the W3C test.';
             $phpcsFile->addError($error, 0, '');
         }
-        if ( ! self::_checkUtf8Rfc3629($file_content)) {
+        if (!self::_checkUtf8Rfc3629($file_content)) {
             $error = 'File "' . $file_name . '" should be saved with Unicode (UTF-8) encoding, but it did not meet RFC3629 requirements.';
             $phpcsFile->addError($error, 0, '');
         }
     }//end process()
-
 
     /**
      * Checks that the string $content contains only valid UTF-8 chars
@@ -91,16 +86,14 @@ class Utf8EncodingSniff implements Sniff
      * @param string $content String to check.
      *
      * @return bool true if $content contains only UTF-8 chars, false otherwise.
-     *
      * @see http://w3.org/International/questions/qa-forms-utf-8.html
      */
     private static function _checkUtf8W3c($content)
     {
-        $content_chunks=self::mb_chunk_split($content, 4096, '');
-    	foreach($content_chunks as $content_chunk)
-		{
-			$preg_result= preg_match(
-            '%^(?:
+        $content_chunks = self::mb_chunk_split($content, 4096, '');
+        foreach ($content_chunks as $content_chunk) {
+            $preg_result = preg_match(
+                '%^(?:
                   [\x09\x0A\x0D\x20-\x7E]            # ASCII
                 | [\xC2-\xDF][\x80-\xBF]             # non-overlong 2-byte
                 | \xE0[\xA0-\xBF][\x80-\xBF]         # excluding overlongs
@@ -110,16 +103,70 @@ class Utf8EncodingSniff implements Sniff
                 | [\xF1-\xF3][\x80-\xBF]{3}          # planes 4-15
                 | \xF4[\x80-\x8F][\x80-\xBF]{2}      # plane 16
             )*$%xs',
-            $content_chunk
-			);
-			if($preg_result!==1)
-			{
-				return false;
-			}
+                $content_chunk
+            );
+            if ($preg_result !== 1) {
+                return false;
+            }
+        }
 
-		}
-		return true;
+        return true;
     }//end _checkUtf8W3c()
+
+    /**
+     * Splits a string to chunks of given size
+     * This helps to avoid segmentation fault errors when large text is given
+     * Returns array of strings after splitting
+     *
+     * @param string $str String to split.
+     * @param int    $len number of characters per chunk
+     * @param string $glue
+     *
+     * @return array string array after splitting
+     * @see http://php.net/manual/en/function.chunk-split.php
+     */
+    private static function mb_chunk_split($str, $len, $glue)
+    {
+        if (empty($str)) {
+            return false;
+        }
+        $array = self::mbStringToArray($str);
+        $n     = -1;
+        $new   = [];
+        foreach ($array as $char) {
+            $n++;
+            if ($n < $len) {
+                $new [] = $char;
+            } elseif ($n == $len) {
+                $new [] = $glue . $char;
+                $n      = 0;
+            }
+        }
+
+        return $new;
+    }//_checkUtf8Rfc3629()
+
+    /**
+     * Supporting function for mb_chunk_split
+     *
+     * @param string $str
+     *
+     * @return array
+     * @see http://php.net/manual/en/function.chunk-split.php
+     */
+    private static function mbStringToArray($str)
+    {
+        if (empty($str)) {
+            return false;
+        }
+        $len   = mb_strlen($str);
+        $array = [];
+        for ($i = 0; $i < $len; $i++) {
+            $array[] = mb_substr($str, $i, 1);
+        }
+
+        return $array;
+    }//mb_chunk_split
 
     /**
      * Checks that the string $content contains only valid UTF-8 chars
@@ -129,7 +176,6 @@ class Utf8EncodingSniff implements Sniff
      * @param string $content String to check.
      *
      * @return bool true if $content contains only UTF-8 chars, false otherwise.
-     *
      * @see http://www.php.net/manual/en/function.mb-detect-encoding.php#85294
      */
     private static function _checkUtf8Rfc3629($content)
@@ -141,9 +187,9 @@ class Utf8EncodingSniff implements Sniff
                 if (($c >= 254)) {
                     return false;
                 } elseif ($c >= 252) {
-                    $bits=6;
+                    $bits = 6;
                 } elseif ($c >= 248) {
-                    $bits=5;
+                    $bits = 5;
                 } elseif ($c >= 240) {
                     $bytes = 4;
                 } elseif ($c >= 224) {
@@ -152,9 +198,11 @@ class Utf8EncodingSniff implements Sniff
                     $bytes = 2;
                 } else {
                     return false;
-                } if (($i + $bytes) > $len) {
+                }
+                if (($i + $bytes) > $len) {
                     return false;
-                } while ($bytes > 1) {
+                }
+                while ($bytes > 1) {
                     $i++;
                     $b = ord($content[$i]);
                     if ($b < 128 || $b > 191) {
@@ -164,59 +212,7 @@ class Utf8EncodingSniff implements Sniff
                 }
             }
         }
+
         return true;
-    }//_checkUtf8Rfc3629()
-	 
-	 /**
-     * Splits a string to chunks of given size
-	 * This helps to avoid segmentation fault errors when large text is given
-     * Returns array of strings after splitting
-     *
-     * @param string $str String to split.
-	 * @param int $len number of characters per chunk
-     *
-     * @return array string array after splitting
-     *
-     * @see http://php.net/manual/en/function.chunk-split.php
-     */
-	private static function mb_chunk_split($str, $len, $glue) 
-	{
-		if (empty($str)) return false;
-		$array = self::mbStringToArray ($str);
-		$n = -1;
-		$new = Array();
-		foreach ($array as $char) {
-			$n++;
-			if ($n < $len) $new []= $char;
-			elseif ($n == $len) {
-				$new []= $glue . $char;
-				$n = 0;
-			}
-		}
-		return $new;
-	}//mb_chunk_split
-	/**
-     * Supporting function for mb_chunk_split
-     *
-     * @param string $str   
-	 *
-     * @return array 
-     *
-     * @see http://php.net/manual/en/function.chunk-split.php
-     */
-	private static function mbStringToArray ($str) 
-	{
-		if (empty($str)) return false;
-		$len = mb_strlen($str);
-		$array = array();
-		for ($i = 0; $i < $len; $i++) {
-			$array[] = mb_substr($str, $i, 1);
-		}
-		return $array;
-	}
-
-	
-
+    }
 }//end class
-
-?>
